+++ projects/graphql-client/src/index.ts
@@ -36,3 +36,4 @@ export * from "./remote/mutations/useCreateShare"
 export * from "./remote/mutations/useCreatePlaylist"
 export * from "./remote/mutations/useChangePassword"
 export * from "./remote/mutations/useAddSongsToPlaylist"
+export * from "./remote/mutations/useAcceptInvitation"
